+++ clean_python/__init__.py
@@ -3,5 +3,5 @@
 from .base import *  # NOQA
 
 # fmt: off
-__version__ = '0.1.3.dev0'
+__version__ = '0.2.0b4.dev0'
 # fmt: on