+++ clean_python/base/domain/domain_service.py
@@ -1,11 +1,10 @@
 # (c) Nelen & Schuurmans
 
 from pydantic import BaseModel
+from pydantic import ConfigDict
 
 __all__ = ["DomainService"]
 
 
 class DomainService(BaseModel):
-    class Config:
-        allow_mutation = False
-        arbitrary_types_allowed = True
+    model_config = ConfigDict(frozen=True, arbitrary_types_allowed=True)