+++ clean_python/fastapi/request_query.py
@@ -3,7 +3,7 @@
 from typing import List
 
 from fastapi import Query
-from pydantic import validator
+from pydantic import field_validator
 
 from clean_python import Filter
 from clean_python import PageOptions
@@ -19,11 +19,11 @@ class RequestQuery(ValueObject):
         default="id", enum=["id", "-id"], description="Field to order by"
     )
 
-    @validator("order_by")
-    def validate_order_by_enum(cls, v):
+    @field_validator("order_by")
+    def validate_order_by_enum(cls, v, _):
         # the 'enum' parameter doesn't actually do anthing in validation
         # See: https://github.com/tiangolo/fastapi/issues/2910
-        allowed = cls.__fields__["order_by"].field_info.extra["enum"]
+        allowed = cls.model_fields["order_by"].json_schema_extra["enum"]
         if v not in allowed:
             raise ValueError(f"'order_by' must be one of {allowed}")
         return v
@@ -41,7 +41,7 @@ class RequestQuery(ValueObject):
 
     def filters(self) -> List[Filter]:
         result = []
-        for name in self.__fields__:
+        for name in self.model_fields:
             if name in {"limit", "offset", "order_by"}:
                 continue
             value = getattr(self, name)