+++ clean_python/oauth2/oauth2.py
@@ -1,7 +1,9 @@
 # (c) Nelen & Schuurmans
 
 from typing import Dict
+from typing import FrozenSet
 from typing import List
+from typing import Optional
 
 import jwt
 from jwt import PyJWKClient
@@ -9,23 +11,30 @@ from jwt.exceptions import PyJWTError
 from pydantic import AnyHttpUrl
 from pydantic import BaseModel
 
-from clean_python.base.domain.exceptions import PermissionDenied
-from clean_python.base.domain.exceptions import Unauthorized
+from clean_python import PermissionDenied
+from clean_python import Unauthorized
 
-__all__ = ["OAuth2Settings", "OAuth2AccessTokenVerifier"]
+from .claims import Claims
+from .claims import Tenant
+
+__all__ = ["TokenVerifier", "TokenVerifierSettings", "OAuth2SPAClientSettings"]
 
 
-class OAuth2Settings(BaseModel):
-    client_id: str
+class TokenVerifierSettings(BaseModel):
     issuer: str
-    resource_server_id: str
+    algorithms: List[str] = ["RS256"]
+    # optional additional checks:
+    scope: Optional[str] = None
+    admin_users: Optional[List[str]] = None  # 'sub' whitelist
+
+
+class OAuth2SPAClientSettings(BaseModel):
+    client_id: str
     token_url: AnyHttpUrl
     authorization_url: AnyHttpUrl
-    algorithms: List[str] = ["RS256"]
-    admin_users: List[str]
 
 
-class OAuth2AccessTokenVerifier:
+class TokenVerifier:
     """A class for verifying OAuth2 Access Tokens from AWS Cognito
 
     The verification steps followed are documented here:
@@ -37,22 +46,20 @@ class OAuth2AccessTokenVerifier:
     # allow 2 minutes leeway for verifying token expiry:
     LEEWAY = 120
 
-    def __init__(
-        self,
-        scope: str,
-        issuer: str,
-        resource_server_id: str,
-        algorithms: List[str],
-        admin_users: List[str],
-    ):
-        self.scope = scope
-        self.issuer = issuer
-        self.algorithms = algorithms
-        self.resource_server_id = resource_server_id
-        self.admin_users = admin_users
-        self.jwk_client = PyJWKClient(f"{issuer}/.well-known/jwks.json")
+    def __init__(self, settings: TokenVerifierSettings):
+        self.settings = settings
+        self.jwk_client = PyJWKClient(f"{settings.issuer}/.well-known/jwks.json")
 
-    def __call__(self, token: str) -> Dict:
+    def __call__(self, authorization: Optional[str]) -> Claims:
+        # Step 0: retrieve the token from the Authorization header
+        # See https://tools.ietf.org/html/rfc6750#section-2.1,
+        # Bearer is case-sensitive and there is exactly 1 separator after.
+        if authorization is None:
+            raise Unauthorized()
+        token = authorization[7:] if authorization.startswith("Bearer") else None
+        if token is None:
+            raise Unauthorized()
+
         # Step 1: Confirm the structure of the JWT. This check is part of get_kid since
         # jwt.get_unverified_header will raise a JWTError if the structure is wrong.
         try:
@@ -65,8 +72,8 @@ class OAuth2AccessTokenVerifier:
             claims = jwt.decode(
                 token,
                 key.key,
-                algorithms=self.algorithms,
-                issuer=self.issuer,
+                algorithms=self.settings.algorithms,
+                issuer=self.settings.issuer,
                 leeway=self.LEEWAY,
                 options={
                     "require": ["exp", "iss", "sub", "scope", "token_use"],
@@ -77,36 +84,46 @@ class OAuth2AccessTokenVerifier:
             raise Unauthorized()
         # Step 3: Verify additional claims. At this point, we have passed
         # verification, so unverified claims may be used safely.
+        scope = self.parse_scope(claims)
+        tenant = self.parse_tenant(claims)
         self.verify_token_use(claims)
-        self.verify_scope(claims)
-        # Step 4: Authorization: we currently work with a hardcoded
-        # list of users ('sub' claims)
-        self.authorize(claims)
-        return claims
+        self.verify_scope(scope)
+        # Step 4: Authorization: verify 'sub' claim against 'admin_users'
+        self.verify_sub(claims)
+        return Claims(scope=scope, tenant=tenant)
 
     def get_key(self, token) -> jwt.PyJWK:
         """Return the JSON Web KEY (JWK) corresponding to kid."""
         return self.jwk_client.get_signing_key_from_jwt(token)
 
-    def verify_token_use(self, claims):
+    def verify_token_use(self, claims: Dict) -> None:
         """Check the token_use claim."""
         if claims["token_use"] != "access":
             # logger.info("Token has invalid token_use claim: %s", claims["token_use"])
             raise Unauthorized()
 
-    def verify_scope(self, claims):
-        """Check scope claim.
-
-        Cognito includes the resource server id inside the scope, like this:
-
-           raster.lizard.net/*.readwrite
-        """
-        if f"{self.resource_server_id}{self.scope}" not in claims["scope"].split(" "):
+    def verify_scope(self, claims_scope: FrozenSet[str]) -> None:
+        """Parse scopes and optionally check scope claim."""
+        if self.settings.scope is None:
+            return
+        if self.settings.scope not in claims_scope:
             # logger.info("Token has invalid scope claim: %s", claims["scope"])
             raise Unauthorized()
 
-    def authorize(self, claims):
+    def verify_sub(self, claims: Dict) -> None:
         """The subject (sub) claim should be in a hard-coded whitelist."""
-        if claims.get("sub") not in self.admin_users:
+        if self.settings.admin_users is None:
+            return
+        if claims.get("sub") not in self.settings.admin_users:
             # logger.info("User with sub %s is not authorized", claims.get("sub"))
             raise PermissionDenied()
+
+    def parse_scope(self, claims: Dict) -> FrozenSet[str]:
+        return frozenset(claims["scope"].split(" "))
+
+    def parse_tenant(self, claims: Dict) -> Optional[Tenant]:
+        if claims.get("tenant"):
+            tenant = Tenant(id=claims["tenant"], name=claims.get("tenant_name", ""))
+        else:
+            tenant = None
+        return tenant